addManualMappingInvoice
